+++ priv/endpoints.exs
@@ -1,3 +1,52 @@
+# Note that these lists are the CONTROL regions (since those are where we can hit the API),
+# not the MEDIA regions.
+# See https://docs.aws.amazon.com/general/latest/gr/chime-sdk.htm
+chime_identity_regions = [
+  "eu-central-1",
+  "us-east-1"
+]
+
+chime_meeting_regions = [
+  "ap-south-1",
+  "ap-northeast-1",
+  "ap-northeast-2",
+  "ap-southeast-1",
+  "ap-southeast-2",
+  "ca-central-1",
+  "eu-central-1",
+  "eu-west-2",
+  "il-central-1",
+  "us-gov-east-1",
+  "us-gov-west-1",
+  "us-east-1",
+  "us-west-2"
+]
+
+chime_media_pipeline_regions = [
+  "ap-southeast-1",
+  "eu-central-1",
+  "us-east-1",
+  "us-west-2"
+]
+
+chime_messaging_regions = [
+  "eu-central-1",
+  "us-east-1"
+]
+
+chime_voice_regions = [
+  "ap-nottheast-1",
+  "ap-nottheast-2",
+  "ap-southeast-1",
+  "ap-southeast-2",
+  "ca-central-1",
+  "eu-central-1",
+  "eu-west-1",
+  "eu-west-2",
+  "us-east-1",
+  "us-west-2"
+]
+
 %{
   "partitions" => [
     %{
@@ -19,7 +68,8 @@
         "ap-southeast-1" => %{"description" => "Asia Pacific (Singapore)"},
         "ap-southeast-2" => %{"description" => "Asia Pacific (Sydney)"},
         "ap-southeast-3" => %{"description" => "Asia Pacific (Jakarta)"},
-        "ca-central-1" => %{"description" => "Canada (Central)"},
+        "ca-central-1" => %{"description" => "Canada (Montreal)"},
+        "ca-west-1" => %{"description" => "Canada (Calgary)"},
         "eu-central-1" => %{"description" => "EU (Frankfurt)"},
         "eu-west-1" => %{"description" => "EU (Ireland)"},
         "eu-west-2" => %{"description" => "EU (London)"},
@@ -34,6 +84,70 @@
         "us-west-2" => %{"description" => "US West (Oregon)"}
       },
       "services" => %{
+        "personalize" => %{
+          "endpoints" => %{
+            "us-east-1" => %{},
+            "us-east-2" => %{},
+            "us-west-2" => %{},
+            "ap-northeast-1" => %{},
+            "ap-northeast-2" => %{},
+            "ap-south-1" => %{},
+            "ap-southeast-1" => %{},
+            "ap-southeast-2" => %{},
+            "ca-central-1" => %{},
+            "eu-west-1" => %{},
+            "eu-central-1" => %{}
+          }
+        },
+        "personalize-events" => %{
+          "endpoints" => %{
+            "us-east-1" => %{},
+            "us-east-2" => %{},
+            "us-west-2" => %{},
+            "ap-northeast-1" => %{},
+            "ap-northeast-2" => %{},
+            "ap-south-1" => %{},
+            "ap-southeast-1" => %{},
+            "ap-southeast-2" => %{},
+            "ca-central-1" => %{},
+            "eu-west-1" => %{},
+            "eu-central-1" => %{}
+          }
+        },
+        "personalize-runtime" => %{
+          "endpoints" => %{
+            "us-east-1" => %{},
+            "us-east-2" => %{},
+            "us-west-2" => %{},
+            "ap-northeast-1" => %{},
+            "ap-northeast-2" => %{},
+            "ap-south-1" => %{},
+            "ap-southeast-1" => %{},
+            "ap-southeast-2" => %{},
+            "ca-central-1" => %{},
+            "eu-west-1" => %{},
+            "eu-central-1" => %{}
+          }
+        },
+        "bedrock" => %{
+          "endpoints" => %{
+            "ap-northeast-1" => %{},
+            "ap-southeast-1" => %{},
+            "eu-central-1" => %{},
+            "us-east-1" => %{},
+            "us-west-2" => %{}
+          }
+        },
+        "bedrock-runtime" => %{
+          "defaults" => %{"credentialScope" => %{"service" => "bedrock"}},
+          "endpoints" => %{
+            "ap-northeast-1" => %{},
+            "ap-southeast-1" => %{},
+            "eu-central-1" => %{},
+            "us-east-1" => %{},
+            "us-west-2" => %{}
+          }
+        },
         "connect" => %{
           "endpoints" => %{
             "ap-northeast-1" => %{},
@@ -55,6 +169,7 @@
             "ca-central-1" => %{},
             "eu-central-1" => %{},
             "eu-west-1" => %{},
+            "eu-west-2" => %{},
             "eu-west-3" => %{},
             "us-east-1" => %{},
             "us-east-2" => %{},
@@ -101,6 +216,7 @@
             "ap-southeast-1" => %{},
             "ap-southeast-2" => %{},
             "ca-central-1" => %{},
+            "ca-west-1" => %{},
             "eu-central-1" => %{},
             "eu-west-1" => %{},
             "eu-west-2" => %{},
@@ -123,6 +239,7 @@
             "ap-southeast-1" => %{},
             "ap-southeast-2" => %{},
             "ca-central-1" => %{},
+            "ca-west-1" => %{},
             "eu-central-1" => %{},
             "eu-west-1" => %{},
             "eu-west-2" => %{},
@@ -135,9 +252,62 @@
             "us-west-2" => %{}
           }
         },
+        "mobiletargeting" => %{
+          "endpoints" => %{
+            "us-east-1" => %{
+              "hostname" => "pinpoint.us-east-1.amazonaws.com"
+            },
+            "us-east-2" => %{
+              "hostname" => "pinpoint.us-east-2.amazonaws.com"
+            },
+            "us-west-2" => %{
+              "hostname" => "pinpoint.us-west-1.amazonaws.com"
+            },
+            "ap-south-1" => %{
+              "hostname" => "pinpoint.ap-south-1.amazonaws.com"
+            },
+            "ap-northeast-1" => %{
+              "hostname" => "pinpoint.ap-northeast-1.amazonaws.com"
+            },
+            "ap-northeast-2" => %{
+              "hostname" => "pinpoint.ap-northeast-2.amazonaws.com"
+            },
+            "ap-southeast-1" => %{
+              "hostname" => "pinpoint.ap-southeast-1.amazonaws.com"
+            },
+            "ap-southeast-2" => %{
+              "hostname" => "pinpoint.ap-southeast-2.amazonaws.com"
+            },
+            "ca-central-1" => %{
+              "hostname" => "pinpoint.ca-central-1.amazonaws.com"
+            },
+            "eu-central-1" => %{
+              "hostname" => "pinpoint.eu-central-1.amazonaws.com"
+            },
+            "eu-west-1" => %{
+              "hostname" => "pinpoint.eu-west-1.amazonaws.com"
+            },
+            "eu-west-2" => %{
+              "hostname" => "pinpoint.eu-west-2.amazonaws.com"
+            }
+          }
+        },
         "pinpoint" => %{
           "defaults" => %{"credentialScope" => %{"service" => "mobiletargeting"}},
-          "endpoints" => %{"us-east-1" => %{}}
+          "endpoints" => %{
+            "us-east-1" => %{},
+            "us-east-2" => %{},
+            "us-west-2" => %{},
+            "ap-south-1" => %{},
+            "ap-northeast-1" => %{},
+            "ap-northeast-2" => %{},
+            "ap-southeast-1" => %{},
+            "ap-southeast-2" => %{},
+            "ca-central-1" => %{},
+            "eu-central-1" => %{},
+            "eu-west-1" => %{},
+            "eu-west-2" => %{}
+          }
         },
         "iam" => %{
           "endpoints" => %{
@@ -242,6 +412,7 @@
             "ap-southeast-1" => %{},
             "ap-southeast-2" => %{},
             "ca-central-1" => %{},
+            "ca-west-1" => %{},
             "eu-central-1" => %{},
             "eu-west-1" => %{},
             "eu-west-2" => %{},
@@ -267,6 +438,7 @@
             "ap-southeast-1" => %{},
             "ap-southeast-2" => %{},
             "ca-central-1" => %{},
+            "ca-west-1" => %{},
             "eu-central-1" => %{"sslCommonName" => "{service}.{region}.{dnsSuffix}"},
             "eu-west-1" => %{},
             "eu-west-2" => %{},
@@ -321,6 +493,7 @@
             "ap-southeast-1" => %{},
             "ap-southeast-2" => %{},
             "ca-central-1" => %{},
+            "ca-west-1" => %{},
             "eu-central-1" => %{},
             "eu-west-1" => %{},
             "eu-west-2" => %{},
@@ -362,6 +535,29 @@
             "ap-southeast-1" => %{},
             "ap-southeast-2" => %{},
             "ca-central-1" => %{},
+            "ca-west-1" => %{},
+            "eu-central-1" => %{},
+            "eu-west-1" => %{},
+            "eu-west-2" => %{},
+            "eu-west-3" => %{},
+            "eu-north-1" => %{},
+            "sa-east-1" => %{},
+            "us-east-1" => %{},
+            "us-east-2" => %{},
+            "us-west-1" => %{},
+            "us-west-2" => %{}
+          }
+        },
+        "scheduler" => %{
+          "endpoints" => %{
+            "ap-northeast-1" => %{},
+            "ap-northeast-2" => %{},
+            "ap-east-1" => %{},
+            "ap-south-1" => %{},
+            "ap-southeast-1" => %{},
+            "ap-southeast-2" => %{},
+            "ca-central-1" => %{},
+            "ca-west-1" => %{},
             "eu-central-1" => %{},
             "eu-west-1" => %{},
             "eu-west-2" => %{},
@@ -383,6 +579,7 @@
             "ap-southeast-1" => %{},
             "ap-southeast-2" => %{},
             "ca-central-1" => %{},
+            "ca-west-1" => %{},
             "eu-central-1" => %{},
             "eu-west-1" => %{},
             "eu-west-2" => %{},
@@ -428,11 +625,13 @@
             "ap-southeast-1" => %{},
             "ap-southeast-2" => %{},
             "ca-central-1" => %{},
+            "ca-west-1" => %{},
             "eu-central-1" => %{},
             "eu-west-1" => %{},
             "eu-west-2" => %{},
             "eu-west-3" => %{},
             "eu-north-1" => %{},
+            "me-south-1" => %{},
             "sa-east-1" => %{},
             "us-east-1" => %{},
             "us-east-2" => %{},
@@ -449,6 +648,8 @@
             "ap-south-1" => %{},
             "ap-southeast-1" => %{},
             "ap-southeast-2" => %{},
+            "ca-central-1" => %{},
+            "ca-west-1" => %{},
             "eu-central-1" => %{},
             "eu-west-1" => %{},
             "eu-west-2" => %{},
@@ -504,6 +705,7 @@
             "ap-southeast-1" => %{},
             "ap-southeast-2" => %{},
             "ca-central-1" => %{},
+            "ca-west-1" => %{},
             "eu-central-1" => %{},
             "eu-west-1" => %{},
             "eu-west-2" => %{},
@@ -526,6 +728,7 @@
             "ap-southeast-1" => %{},
             "ap-southeast-2" => %{},
             "ca-central-1" => %{},
+            "ca-west-1" => %{},
             "eu-central-1" => %{},
             "eu-west-1" => %{},
             "eu-west-2" => %{},
@@ -555,6 +758,7 @@
             "ap-southeast-1" => %{},
             "ap-southeast-2" => %{},
             "ca-central-1" => %{},
+            "ca-west-1" => %{},
             "eu-central-1" => %{},
             "eu-west-1" => %{},
             "eu-west-2" => %{},
@@ -585,6 +789,29 @@
             "ap-southeast-1" => %{},
             "ap-southeast-2" => %{},
             "ca-central-1" => %{},
+            "ca-west-1" => %{},
+            "eu-central-1" => %{},
+            "eu-west-1" => %{},
+            "eu-west-2" => %{},
+            "eu-west-3" => %{},
+            "eu-north-1" => %{},
+            "sa-east-1" => %{},
+            "us-east-1" => %{},
+            "us-east-2" => %{},
+            "us-west-1" => %{},
+            "us-west-2" => %{}
+          }
+        },
+        "aoss" => %{
+          "endpoints" => %{
+            "ap-northeast-1" => %{},
+            "ap-northeast-2" => %{},
+            "ap-east-1" => %{},
+            "ap-south-1" => %{},
+            "ap-southeast-1" => %{},
+            "ap-southeast-2" => %{},
+            "ca-central-1" => %{},
+            "ca-west-1" => %{},
             "eu-central-1" => %{},
             "eu-west-1" => %{},
             "eu-west-2" => %{},
@@ -674,6 +901,7 @@
             "ap-southeast-1" => %{},
             "ap-southeast-2" => %{},
             "ca-central-1" => %{},
+            "ca-west-1" => %{},
             "eu-central-1" => %{},
             "eu-west-1" => %{},
             "eu-west-2" => %{},
@@ -697,6 +925,7 @@
           "endpoints" => %{
             "ap-northeast-1" => %{},
             "eu-west-1" => %{},
+            "eu-west-2" => %{},
             "us-east-1" => %{},
             "us-east-2" => %{},
             "us-west-2" => %{}
@@ -790,6 +1019,29 @@
             "us-west-2" => %{}
           }
         },
+        "ivs" => %{
+          "defaults" => %{"protocols" => ["http", "https"]},
+          "endpoints" => %{
+            "ap-northeast-1" => %{},
+            "ap-northeast-2" => %{},
+            "ap-east-1" => %{},
+            "ap-south-1" => %{},
+            "ap-southeast-1" => %{},
+            "ap-southeast-2" => %{},
+            "ca-central-1" => %{},
+            "eu-central-1" => %{},
+            "eu-west-1" => %{},
+            "eu-west-2" => %{},
+            "eu-west-3" => %{},
+            "eu-north-1" => %{},
+            "eu-south-1" => %{},
+            "sa-east-1" => %{},
+            "us-east-1" => %{},
+            "us-east-2" => %{},
+            "us-west-1" => %{},
+            "us-west-2" => %{}
+          }
+        },
         "ec2" => %{
           "defaults" => %{"protocols" => ["http", "https"]},
           "endpoints" => %{
@@ -801,6 +1053,7 @@
             "ap-southeast-1" => %{},
             "ap-southeast-2" => %{},
             "ca-central-1" => %{},
+            "ca-west-1" => %{},
             "eu-central-1" => %{},
             "eu-west-1" => %{},
             "eu-west-2" => %{},
@@ -850,6 +1103,16 @@
           "isRegionalized" => false,
           "partitionEndpoint" => "aws-global"
         },
+        "bcm-data-exports" => %{
+          "endpoints" => %{
+            "aws-global" => %{
+              "credentialScope" => %{"region" => "us-east-1"},
+              "hostname" => "bcm-data-exports.us-east-1.api.aws"
+            }
+          },
+          "isRegionalized" => false,
+          "partitionEndpoint" => "aws-global"
+        },
         "ds" => %{
           "endpoints" => %{
             "ap-northeast-1" => %{},
@@ -858,6 +1121,7 @@
             "ap-southeast-1" => %{},
             "ap-southeast-2" => %{},
             "ca-central-1" => %{},
+            "ca-west-1" => %{},
             "eu-central-1" => %{},
             "eu-west-1" => %{},
             "eu-west-2" => %{},
@@ -892,6 +1156,7 @@
             "ap-southeast-1" => %{},
             "ap-southeast-2" => %{},
             "ca-central-1" => %{},
+            "ca-west-1" => %{},
             "eu-central-1" => %{},
             "eu-west-1" => %{},
             "eu-west-2" => %{},
@@ -916,20 +1181,31 @@
         },
         "kms" => %{
           "endpoints" => %{
+            "af-south-1" => %{},
             "ap-northeast-1" => %{},
             "ap-northeast-2" => %{},
             "ap-east-1" => %{},
             "ap-south-1" => %{},
             "ap-southeast-1" => %{},
             "ap-southeast-2" => %{},
+            "ap-southeast-3" => %{},
             "ca-central-1" => %{},
+            "ca-west-1" => %{},
             "eu-central-1" => %{},
+            "eu-central-2" => %{},
+            "eu-south-1" => %{},
             "eu-west-1" => %{},
             "eu-west-2" => %{},
             "eu-west-3" => %{},
             "eu-north-1" => %{},
+            "me-central-1" => %{},
+            "me-south-1" => %{},
             "sa-east-1" => %{},
             "us-east-1" => %{},
+            "us-east-1-fips" => %{
+              "hostname" => "kms-fips.us-east-1.amazonaws.com",
+              "credentialScope" => %{"region" => "us-east-1"}
+            },
             "us-east-2" => %{},
             "us-west-1" => %{},
             "us-west-2" => %{}
@@ -980,14 +1256,25 @@
             "ap-south-1" => %{},
             "ap-southeast-1" => %{},
             "ap-southeast-2" => %{},
+            "ap-southeast-3" => %{},
             "ca-central-1" => %{},
+            "ca-west-1" => %{},
             "eu-central-1" => %{},
+            "eu-central-2" => %{},
+            "eu-south-1" => %{},
             "eu-west-1" => %{},
             "eu-west-2" => %{},
             "eu-west-3" => %{},
             "eu-north-1" => %{},
+            "me-central-1" => %{},
+            "me-south-1" => %{},
             "sa-east-1" => %{},
             "us-east-1" => %{"sslCommonName" => "{service}.{dnsSuffix}"},
+            "us-east-1-fips" => %{
+              "sslCommonName" => "{service}.{dnsSuffix}",
+              "hostname" => "rds-fips.us-east-1.amazonaws.com",
+              "credentialScope" => %{"region" => "us-east-1"}
+            },
             "us-east-2" => %{},
             "us-west-1" => %{},
             "us-west-2" => %{}
@@ -1064,6 +1351,7 @@
             "ap-southeast-1" => %{},
             "ap-southeast-2" => %{},
             "ca-central-1" => %{},
+            "ca-west-1" => %{},
             "eu-central-1" => %{},
             "eu-west-1" => %{},
             "eu-west-2" => %{},
@@ -1118,6 +1406,49 @@
             "ap-southeast-2" => %{},
             "ca-central-1" => %{},
             "eu-central-1" => %{},
+            "eu-central-2" => %{},
+            "eu-west-1" => %{},
+            "eu-west-2" => %{},
+            "eu-west-3" => %{},
+            "eu-north-1" => %{},
+            "sa-east-1" => %{},
+            "us-east-1" => %{},
+            "us-east-2" => %{},
+            "us-west-1" => %{},
+            "us-west-2" => %{}
+          }
+        },
+        "ssm-incidents" => %{
+          "endpoints" => %{
+            "ap-northeast-1" => %{},
+            "ap-northeast-2" => %{},
+            "ap-east-1" => %{},
+            "ap-south-1" => %{},
+            "ap-southeast-1" => %{},
+            "ap-southeast-2" => %{},
+            "ca-central-1" => %{},
+            "eu-central-1" => %{},
+            "eu-west-1" => %{},
+            "eu-west-2" => %{},
+            "eu-west-3" => %{},
+            "eu-north-1" => %{},
+            "sa-east-1" => %{},
+            "us-east-1" => %{},
+            "us-east-2" => %{},
+            "us-west-1" => %{},
+            "us-west-2" => %{}
+          }
+        },
+        "ssm-contacts" => %{
+          "endpoints" => %{
+            "ap-northeast-1" => %{},
+            "ap-northeast-2" => %{},
+            "ap-east-1" => %{},
+            "ap-south-1" => %{},
+            "ap-southeast-1" => %{},
+            "ap-southeast-2" => %{},
+            "ca-central-1" => %{},
+            "eu-central-1" => %{},
             "eu-west-1" => %{},
             "eu-west-2" => %{},
             "eu-west-3" => %{},
@@ -1137,6 +1468,7 @@
             "ap-southeast-1" => %{},
             "ap-southeast-2" => %{},
             "ca-central-1" => %{},
+            "ca-west-1" => %{},
             "eu-central-1" => %{},
             "eu-west-1" => %{},
             "eu-west-2" => %{},
@@ -1171,6 +1503,7 @@
             "ap-southeast-1" => %{},
             "ap-southeast-2" => %{},
             "ca-central-1" => %{},
+            "ca-west-1" => %{},
             "eu-central-1" => %{},
             "eu-west-1" => %{},
             "eu-west-2" => %{},
@@ -1192,6 +1525,7 @@
             "ap-southeast-1" => %{},
             "ap-southeast-2" => %{},
             "ca-central-1" => %{},
+            "ca-west-1" => %{},
             "eu-central-1" => %{},
             "eu-west-1" => %{},
             "eu-west-2" => %{},
@@ -1212,6 +1546,7 @@
             "ap-southeast-1" => %{},
             "ap-southeast-2" => %{},
             "ca-central-1" => %{},
+            "ca-west-1" => %{},
             "eu-central-1" => %{},
             "eu-west-1" => %{},
             "eu-west-2" => %{},
@@ -1321,6 +1656,7 @@
             "us-west-2" => %{}
           }
         },
+        # Deprecated - will not work after 2024-06-29:
         "chime" => %{
           "endpoints" => %{
             "aws-global" => %{
@@ -1330,6 +1666,51 @@
           },
           "isRegionalized" => false,
           "partitionEndpoint" => "aws-global"
+        },
+        "chime-sdk-identity" => %{
+          "defaults" => %{"signatureVersions" => ["v4"]},
+          "endpoints" =>
+            Enum.map(
+              chime_identity_regions,
+              &{&1, %{"hostname" => "identity-chime.#{&1}.amazonaws.com"}}
+            )
+            |> Map.new()
+        },
+        "chime-sdk-media-pipelines" => %{
+          "defaults" => %{"signatureVersions" => ["v4"]},
+          "endpoints" =>
+            Enum.map(
+              chime_media_pipeline_regions,
+              &{&1, %{"hostname" => "media-pipelines-chime.#{&1}.amazonaws.com"}}
+            )
+            |> Map.new()
+        },
+        "chime-sdk-meetings" => %{
+          "defaults" => %{"signatureVersions" => ["v4"]},
+          "endpoints" =>
+            Enum.map(
+              chime_meeting_regions,
+              &{&1, %{"hostname" => "meetings-chime.#{&1}.amazonaws.com"}}
+            )
+            |> Map.new()
+        },
+        "chime-sdk-messaging" => %{
+          "defaults" => %{"signatureVersions" => ["v4"]},
+          "endpoints" =>
+            Enum.map(
+              chime_messaging_regions,
+              &{&1, %{"hostname" => "messaging-chime.#{&1}.amazonaws.com"}}
+            )
+            |> Map.new()
+        },
+        "chime-sdk-voice" => %{
+          "defaults" => %{"signatureVersions" => ["v4"]},
+          "endpoints" =>
+            Enum.map(
+              chime_voice_regions,
+              &{&1, %{"hostname" => "voice-chime.#{&1}.amazonaws.com"}}
+            )
+            |> Map.new()
         },
         "s3" => %{
           "defaults" => %{"protocols" => ["http", "https"], "signatureVersions" => ["s3v4"]},
@@ -1358,7 +1739,9 @@
               "signatureVersions" => ["s3", "s3v4"]
             },
             "ca-central-1" => %{},
+            "ca-west-1" => %{},
             "eu-central-1" => %{},
+            "eu-central-2" => %{},
             "eu-west-1" => %{
               "hostname" => "s3.eu-west-1.amazonaws.com",
               "signatureVersions" => ["s3", "s3v4"]
@@ -1367,6 +1750,8 @@
             "eu-west-2" => %{},
             "eu-west-3" => %{},
             "eu-north-1" => %{},
+            "me-central-1" => %{},
+            "me-south-1" => %{},
             "s3-external-1" => %{
               "credentialScope" => %{"region" => "us-east-1"},
               "hostname" => "s3-external-1.amazonaws.com",
@@ -1375,6 +1760,10 @@
             "sa-east-1" => %{
               "hostname" => "s3.sa-east-1.amazonaws.com",
               "signatureVersions" => ["s3", "s3v4"]
+            },
+            "us-east-1-fips" => %{
+              "hostname" => "s3-fips.us-east-1.amazonaws.com",
+              "credentialScope" => %{"region" => "us-east-1"}
             },
             "us-east-1" => %{
               "hostname" => "s3.amazonaws.com",
@@ -1457,6 +1846,7 @@
             "ap-southeast-1" => %{},
             "ap-southeast-2" => %{},
             "ca-central-1" => %{},
+            "ca-west-1" => %{},
             "eu-central-1" => %{},
             "eu-west-1" => %{},
             "eu-west-2" => %{},
@@ -1485,13 +1875,19 @@
             "ap-south-1" => %{},
             "ap-southeast-1" => %{},
             "ap-southeast-2" => %{},
+            "ap-southeast-3" => %{},
             "aws-global" => %{},
             "ca-central-1" => %{},
+            "ca-west-1" => %{},
             "eu-central-1" => %{},
+            "eu-central-2" => %{},
+            "eu-south-1" => %{},
             "eu-west-1" => %{},
             "eu-west-2" => %{},
             "eu-west-3" => %{},
             "eu-north-1" => %{},
+            "me-central-1" => %{},
+            "me-south-1" => %{},
             "sa-east-1" => %{},
             "us-east-1" => %{},
             "us-east-1-fips" => %{
@@ -1522,6 +1918,7 @@
             "sslCommonName" => "{region}.queue.{dnsSuffix}"
           },
           "endpoints" => %{
+            "af-south-1" => %{},
             "ap-northeast-1" => %{},
             "ap-northeast-2" => %{},
             "ap-east-1" => %{},
@@ -1529,11 +1926,14 @@
             "ap-southeast-1" => %{},
             "ap-southeast-2" => %{},
             "ca-central-1" => %{},
+            "ca-west-1" => %{},
             "eu-central-1" => %{},
+            "eu-central-2" => %{},
             "eu-west-1" => %{},
             "eu-west-2" => %{},
             "eu-west-3" => %{},
             "eu-north-1" => %{},
+            "eu-south-1" => %{},
             "fips-us-east-1" => %{},
             "fips-us-east-2" => %{},
             "fips-us-west-1" => %{},
@@ -1570,6 +1970,7 @@
             "us-east-1" => %{},
             "us-west-1" => %{},
             "us-west-2" => %{},
+            "ca-central-1" => %{},
             "ap-southeast-1" => %{},
             "ap-southeast-2" => %{},
             "eu-west-1" => %{},
@@ -1619,6 +2020,7 @@
             "ap-southeast-1" => %{},
             "ap-southeast-2" => %{},
             "ca-central-1" => %{},
+            "ca-west-1" => %{},
             "eu-central-1" => %{},
             "eu-west-1" => %{},
             "eu-west-2" => %{},
@@ -1640,6 +2042,7 @@
             "ap-southeast-1" => %{},
             "ap-southeast-2" => %{},
             "ca-central-1" => %{},
+            "ca-west-1" => %{},
             "eu-central-1" => %{},
             "eu-west-1" => %{},
             "eu-west-2" => %{},
@@ -1683,6 +2086,7 @@
             "ap-southeast-1" => %{},
             "ap-southeast-2" => %{},
             "ca-central-1" => %{},
+            "ca-west-1" => %{},
             "eu-central-1" => %{},
             "eu-west-1" => %{},
             "eu-west-2" => %{},
@@ -1708,6 +2112,7 @@
             "ap-southeast-1" => %{},
             "ap-southeast-2" => %{},
             "ca-central-1" => %{},
+            "ca-west-1" => %{},
             "eu-central-1" => %{},
             "eu-west-1" => %{},
             "eu-west-2" => %{},
@@ -1778,6 +2183,7 @@
             "ap-southeast-1" => %{},
             "ap-southeast-2" => %{},
             "ca-central-1" => %{},
+            "ca-west-1" => %{},
             "eu-central-1" => %{},
             "eu-west-1" => %{},
             "eu-west-2" => %{},
@@ -1806,6 +2212,7 @@
             "ap-southeast-1" => %{},
             "ap-southeast-2" => %{},
             "ca-central-1" => %{},
+            "ca-west-1" => %{},
             "eu-central-1" => %{},
             "eu-west-1" => %{},
             "us-east-1" => %{},
@@ -1813,10 +2220,45 @@
             "us-west-2" => %{}
           }
         },
+        "fsx" => %{
+          "endpoints" => %{
+            "us-east-2" => %{},
+            "us-east-1" => %{},
+            "us-west-1" => %{},
+            "us-west-2" => %{},
+            "af-south-1" => %{},
+            "ap-east-1" => %{},
+            "ap-south-2" => %{},
+            "ap-southeast-3" => %{},
+            "ap-south-1" => %{},
+            "ap-northeast-3" => %{},
+            "ap-northeast-2" => %{},
+            "ap-southeast-1" => %{},
+            "ap-southeast-2" => %{},
+            "ap-northeast-1" => %{},
+            "ca-central-1" => %{},
+            "ca-west-1" => %{},
+            "eu-central-1" => %{},
+            "eu-west-1" => %{},
+            "eu-west-2" => %{},
+            "eu-south-1" => %{},
+            "eu-west-3" => %{},
+            "eu-south-2" => %{},
+            "eu-north-1" => %{},
+            "eu-central-2" => %{},
+            "me-south-1" => %{},
+            "me-central-1" => %{},
+            "sa-east-1" => %{},
+            "us-gov-east-1" => %{},
+            "us-gov-west-1" => %{}
+          }
+        },
         "elasticfilesystem" => %{
           "endpoints" => %{
             "ap-northeast-2" => %{},
             "ap-southeast-2" => %{},
+            "ca-central-1" => %{},
+            "ca-west-1" => %{},
             "eu-central-1" => %{},
             "eu-west-1" => %{},
             "us-east-1" => %{},
@@ -1833,6 +2275,7 @@
             "ap-southeast-1" => %{},
             "ap-southeast-2" => %{},
             "ca-central-1" => %{},
+            "ca-west-1" => %{},
             "eu-central-1" => %{},
             "eu-west-1" => %{},
             "eu-west-2" => %{},
@@ -1882,6 +2325,7 @@
             "ap-southeast-1" => %{},
             "ap-southeast-2" => %{},
             "ca-central-1" => %{},
+            "ca-west-1" => %{},
             "eu-central-1" => %{},
             "eu-west-1" => %{},
             "eu-west-2" => %{},
@@ -1903,6 +2347,7 @@
             "ap-southeast-1" => %{},
             "ap-southeast-2" => %{},
             "ca-central-1" => %{},
+            "ca-west-1" => %{},
             "eu-central-1" => %{},
             "eu-west-1" => %{},
             "eu-west-2" => %{},
@@ -1940,6 +2385,7 @@
             "ap-southeast-1" => %{},
             "ap-southeast-2" => %{},
             "ca-central-1" => %{},
+            "ca-west-1" => %{},
             "eu-central-1" => %{},
             "eu-west-1" => %{},
             "eu-west-2" => %{},
@@ -1958,7 +2404,9 @@
             "ap-southeast-1" => %{},
             "ap-southeast-2" => %{},
             "ca-central-1" => %{},
+            "ca-west-1" => %{},
             "eu-central-1" => %{},
+            "eu-central-2" => %{},
             "eu-west-1" => %{},
             "eu-west-2" => %{},
             "eu-west-3" => %{},
@@ -1993,6 +2441,7 @@
             "ap-southeast-1" => %{},
             "ap-southeast-2" => %{},
             "ca-central-1" => %{},
+            "ca-west-1" => %{},
             "eu-central-1" => %{},
             "eu-west-1" => %{},
             "eu-west-2" => %{},
@@ -2013,6 +2462,7 @@
             "ap-southeast-1" => %{},
             "ap-southeast-2" => %{},
             "ca-central-1" => %{},
+            "ca-west-1" => %{},
             "eu-central-1" => %{},
             "eu-west-1" => %{},
             "eu-west-2" => %{},
@@ -2061,6 +2511,7 @@
             "ap-southeast-1" => %{},
             "ap-southeast-2" => %{},
             "ca-central-1" => %{},
+            "ca-west-1" => %{},
             "eu-central-1" => %{},
             "eu-west-1" => %{},
             "eu-west-2" => %{},
@@ -2080,6 +2531,7 @@
             "ap-southeast-1" => %{},
             "ap-southeast-2" => %{},
             "ca-central-1" => %{},
+            "ca-west-1" => %{},
             "eu-central-1" => %{},
             "eu-west-1" => %{},
             "eu-west-2" => %{},
@@ -2143,6 +2595,7 @@
             "ap-southeast-1" => %{},
             "ap-southeast-2" => %{},
             "ca-central-1" => %{},
+            "ca-west-1" => %{},
             "eu-central-1" => %{},
             "eu-west-1" => %{},
             "eu-west-2" => %{},
@@ -2194,65 +2647,90 @@
           "endpoints" => %{
             "us-east-1" => %{},
             "us-east-2" => %{},
-            "us-west-1" => %{},
+            "us-west-2" => %{},
             "eu-north-1" => %{},
             "eu-central-1" => %{},
             "eu-west-1" => %{},
+            "eu-west-2" => %{},
             "ap-northeast-1" => %{},
+            "ap-south-1" => %{},
             "ap-southeast-1" => %{},
-            "ap-southeast-2" => %{}
+            "ap-southeast-2" => %{},
+            "ca-central-1" => %{},
+            "sa-east-1" => %{},
+            "us-gov-west-1" => %{}
           }
         },
         "maps.geo" => %{
           "endpoints" => %{
             "us-east-1" => %{},
             "us-east-2" => %{},
-            "us-west-1" => %{},
+            "us-west-2" => %{},
             "eu-north-1" => %{},
             "eu-central-1" => %{},
             "eu-west-1" => %{},
+            "eu-west-2" => %{},
             "ap-northeast-1" => %{},
+            "ap-south-1" => %{},
             "ap-southeast-1" => %{},
-            "ap-southeast-2" => %{}
+            "ap-southeast-2" => %{},
+            "ca-central-1" => %{},
+            "sa-east-1" => %{},
+            "us-gov-west-1" => %{}
           }
         },
         "geofencing.geo" => %{
           "endpoints" => %{
             "us-east-1" => %{},
             "us-east-2" => %{},
-            "us-west-1" => %{},
+            "us-west-2" => %{},
             "eu-north-1" => %{},
             "eu-central-1" => %{},
             "eu-west-1" => %{},
+            "eu-west-2" => %{},
             "ap-northeast-1" => %{},
+            "ap-south-1" => %{},
             "ap-southeast-1" => %{},
-            "ap-southeast-2" => %{}
+            "ap-southeast-2" => %{},
+            "ca-central-1" => %{},
+            "sa-east-1" => %{},
+            "us-gov-west-1" => %{}
           }
         },
         "tracking.geo" => %{
           "endpoints" => %{
             "us-east-1" => %{},
             "us-east-2" => %{},
-            "us-west-1" => %{},
+            "us-west-2" => %{},
             "eu-north-1" => %{},
             "eu-central-1" => %{},
             "eu-west-1" => %{},
+            "eu-west-2" => %{},
             "ap-northeast-1" => %{},
+            "ap-south-1" => %{},
             "ap-southeast-1" => %{},
-            "ap-southeast-2" => %{}
+            "ap-southeast-2" => %{},
+            "ca-central-1" => %{},
+            "sa-east-1" => %{},
+            "us-gov-west-1" => %{}
           }
         },
         "routes.geo" => %{
           "endpoints" => %{
             "us-east-1" => %{},
             "us-east-2" => %{},
-            "us-west-1" => %{},
+            "us-west-2" => %{},
             "eu-north-1" => %{},
             "eu-central-1" => %{},
             "eu-west-1" => %{},
+            "eu-west-2" => %{},
             "ap-northeast-1" => %{},
+            "ap-south-1" => %{},
             "ap-southeast-1" => %{},
-            "ap-southeast-2" => %{}
+            "ap-southeast-2" => %{},
+            "ca-central-1" => %{},
+            "sa-east-1" => %{},
+            "us-gov-west-1" => %{}
           }
         },
         "sso" => %{
@@ -2269,6 +2747,31 @@
             "eu-west-2" => %{},
             "eu-west-3" => %{},
             "eu-north-1" => %{},
+            "sa-east-1" => %{},
+            "us-east-1" => %{},
+            "us-east-2" => %{},
+            "us-west-1" => %{},
+            "us-west-2" => %{}
+          }
+        },
+        "pipes" => %{
+          "endpoints" => %{
+            "af-south-1" => %{},
+            "ap-northeast-1" => %{},
+            "ap-northeast-2" => %{},
+            "ap-east-1" => %{},
+            "ap-south-1" => %{},
+            "ap-southeast-1" => %{},
+            "ap-southeast-2" => %{},
+            "ap-southeast-3" => %{},
+            "ca-central-1" => %{},
+            "eu-central-1" => %{},
+            "eu-west-1" => %{},
+            "eu-west-2" => %{},
+            "eu-west-3" => %{},
+            "eu-north-1" => %{},
+            "eu-south-1" => %{},
+            "me-south-1" => %{},
             "sa-east-1" => %{},
             "us-east-1" => %{},
             "us-east-2" => %{},
@@ -2293,6 +2796,9 @@
         "cn-northwest-1" => %{"description" => "China (Ningxia)"}
       },
       "services" => %{
+        "personalize" => %{"endpoints" => %{"cn-north-1" => %{}}},
+        "personalize-runtime" => %{"endpoints" => %{"cn-north-1" => %{}}},
+        "personalize-events" => %{"endpoints" => %{"cn-north-1" => %{}}},
         "elasticache" => %{"endpoints" => %{"cn-north-1" => %{}, "cn-northwest-1" => %{}}},
         "config" => %{"endpoints" => %{"cn-north-1" => %{}, "cn-northwest-1" => %{}}},
         "iam" => %{
@@ -2334,6 +2840,7 @@
           "endpoints" => %{"cn-north-1" => %{}, "cn-northwest-1" => %{}}
         },
         "es" => %{"endpoints" => %{"cn-northwest-1" => %{}}},
+        "aoss" => %{"endpoints" => %{"cn-northwest-1" => %{}}},
         "monitoring" => %{
           "defaults" => %{"protocols" => ["http", "https"]},
           "endpoints" => %{"cn-north-1" => %{}, "cn-northwest-1" => %{}}
@@ -2416,6 +2923,18 @@
       },
       "services" => %{
         "elasticache" => %{"endpoints" => %{"us-gov-east-1" => %{}, "us-gov-west-1" => %{}}},
+        "fsx" => %{
+          "endpoints" => %{
+            "us-gov-east-1" => %{
+              "credentialScope" => %{"region" => "us-gov-east-1"},
+              "hostname" => "fsx-fips.us-gov-east-1.amazonaws.com"
+            },
+            "us-gov-west-1" => %{
+              "credentialScope" => %{"region" => "us-gov-west-1"},
+              "hostname" => "fsx-fips.us-gov-west-1.amazonaws.com"
+            }
+          }
+        },
         "config" => %{"endpoints" => %{"us-gov-east-1" => %{}, "us-gov-west-1" => %{}}},
         "iam" => %{
           "endpoints" => %{
@@ -2482,6 +3001,7 @@
           }
         },
         "es" => %{"endpoints" => %{"us-gov-east-1" => %{}, "us-gov-west-1" => %{}}},
+        "aoss" => %{"endpoints" => %{"us-gov-east-1" => %{}, "us-gov-west-1" => %{}}},
         "monitoring" => %{"endpoints" => %{"us-gov-east-1" => %{}, "us-gov-west-1" => %{}}},
         "elasticloadbalancing" => %{
           "endpoints" => %{
@@ -2495,7 +3015,15 @@
         },
         "ec2" => %{"endpoints" => %{"us-gov-east-1" => %{}, "us-gov-west-1" => %{}}},
         "apigateway" => %{"endpoints" => %{"us-gov-east-1" => %{}, "us-gov-west-1" => %{}}},
-        "kms" => %{"endpoints" => %{"us-gov-east-1" => %{}, "us-gov-west-1" => %{}}},
+        "kms" => %{
+          "endpoints" => %{
+            "us-gov-east-1" => %{
+              "hostname" => "kms-fips.us-gov-east-1.amazonaws.com",
+              "credentialScope" => %{"region" => "us-gov-east-1"}
+            },
+            "us-gov-west-1" => %{}
+          }
+        },
         "sms" => %{"endpoints" => %{"us-gov-east-1" => %{}, "us-gov-west-1" => %{}}},
         "rds" => %{"endpoints" => %{"us-gov-east-1" => %{}, "us-gov-west-1" => %{}}},
         "cloudtrail" => %{"endpoints" => %{"us-gov-east-1" => %{}, "us-gov-west-1" => %{}}},
@@ -2529,11 +3057,11 @@
               "hostname" => "s3-fips.us-gov-west-1.amazonaws.com"
             },
             "us-gov-east-1" => %{
-              "hostname" => "s3.us-gov-east-1.amazonaws.com",
+              "hostname" => "s3-fips.us-gov-east-1.amazonaws.com",
               "protocols" => ["http", "https"]
             },
             "us-gov-west-1" => %{
-              "hostname" => "s3.us-gov-west-1.amazonaws.com",
+              "hostname" => "s3-fips.us-gov-west-1.amazonaws.com",
               "protocols" => ["http", "https"]
             }
           }